Create Channel Islands
Hi, not sure on the format of how to present this but I did my best. I'd like to add the Channel Islands (which should be a seperate entity to the UK). Guernsey and Jersey are the largest islands and have flags shown on 4chan. I'd like to add my regional Jersey flag, Trinity.

Thanks